chore: add development guide and check script
Collect the coding and commiting guides that we want to standardize on.
The check script is a convinience tool to make sure the code follows ours guide,
it can be automated once we enable a CI.

Test coverage related rules are disabled as there are no unit/integration tests
available.